+++ src/main/java/com/tsAdmin/model/Product.java
@@ -5,9 +5,9 @@ public class Product
 {
     private final ProductType type;
     private int quantity;
-    private int volume;
+    private double volume;
 
-    public Product(ProductType type, int quantity, int volume)
+    public Product(ProductType type, int quantity, double volume)
     {
         this.type = type;
         this.quantity = quantity;
@@ -16,10 +16,10 @@ public class Product
 
     // Setter
     public void setQuantity(int quantity) { this.quantity = quantity; }
-    public void setVolume(int volume) { this.volume = volume; }
+    public void setVolume(double volume) { this.volume = volume; }
 
     // Getter
     public ProductType getType() { return type; }
     public int getQuantity() { return quantity; }
-    public int getVolume() { return volume; }
+    public double getVolume() { return volume; }
 }
